open ai usage and onChatCompletion handler
