+++ packages/core/src/effects/generate-message.effects.ts
@@ -1,7 +1,3 @@
-import { s } from '../schema';
-import { DeepPartial } from '../utils';
-import { sleep, switchAsync } from '../utils/async';
-import { createEffect } from '../utils/micro-ngrx';
 import { apiActions, devActions, internalActions } from '../actions';
 import { decodeFrames } from '../frames/decode-frames';
 import { Chat } from '../models';
@@ -18,6 +14,10 @@ import {
   selectShouldGenerateMessage,
   selectSystem,
 } from '../reducers';
+import { s } from '../schema';
+import { sleep, switchAsync } from '../utils/async';
+import { createEffect } from '../utils/micro-ngrx';
+import { updateMessagesWithDelta } from '../utils/update-message';
 
 let currentAbortController: AbortController | null = null;
 
@@ -140,7 +140,7 @@ export const generateMessage = createEffect((store) => {
           })) {
             switch (frame.type) {
               case 'chunk': {
-                message = _updateMessagesWithDelta(message, frame.chunk);
+                message = updateMessagesWithDelta(message, frame.chunk);
                 if (message) {
                   store.dispatch(apiActions.generateMessageChunk(message));
                 }
@@ -203,69 +203,3 @@ export const generateMessage = createEffect((store) => {
     cancelAbortController.abort();
   };
 });
-
-/**
- * Merges existing and new tool calls.
- *
- * @param existingCalls - The existing tool calls.
- * @param newCalls - The new tool calls to merge.
- * @returns The merged array of tool calls.
- */
-function mergeToolCalls(
-  existingCalls: Chat.Api.ToolCall[] = [],
-  newCalls: DeepPartial<Chat.Api.ToolCall>[] = [],
-): Chat.Api.ToolCall[] {
-  const merged = [...existingCalls];
-  newCalls.forEach((newCall) => {
-    const index = merged.findIndex((call) => call.index === newCall.index);
-    if (index !== -1) {
-      const existing = merged[index];
-      merged[index] = {
-        ...existing,
-        function: {
-          ...existing.function,
-          arguments:
-            existing.function.arguments + (newCall.function?.arguments ?? ''),
-        },
-      };
-    } else {
-      merged.push(newCall as Chat.Api.ToolCall);
-    }
-  });
-  return merged;
-}
-
-/**
- * Updates the messages array with an incoming assistant delta.
- *
- * @param messages - The current messages array.
- * @param delta - The incoming message delta.
- * @returns The updated messages array.
- */
-export function _updateMessagesWithDelta(
-  message: Chat.Api.AssistantMessage | null,
-  delta: Chat.Api.CompletionChunk,
-): Chat.Api.AssistantMessage | null {
-  if (message && message.role === 'assistant' && delta.choices.length) {
-    const updatedToolCalls = mergeToolCalls(
-      message.toolCalls,
-      delta.choices[0].delta.toolCalls ?? [],
-    );
-    const updatedMessage: Chat.Api.AssistantMessage = {
-      ...message,
-      content: (message.content ?? '') + (delta.choices[0].delta.content ?? ''),
-      toolCalls: updatedToolCalls,
-    };
-    return updatedMessage;
-  } else if (
-    delta.choices.length &&
-    delta.choices[0]?.delta?.role === 'assistant'
-  ) {
-    return {
-      role: 'assistant',
-      content: delta.choices[0].delta.content ?? '',
-      toolCalls: mergeToolCalls([], delta.choices[0].delta.toolCalls ?? []),
-    };
-  }
-  return message;
-}
