+++ packages/core/src/utils/index.ts
@@ -1,4 +1,5 @@
-export * from './expose-component';
-export * from './types';
-export * from './llm';
 export * from './deep-equal';
+export * from './expose-component';
+export * from './llm';
+export * from './types';
+export { updateMessagesWithDelta } from './update-message';
