+++ samples/server/src/main.ts
@@ -1,12 +1,12 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Chat, KnownModelIds } from '@hashbrownai/core';
-import { HashbrownAzure } from '@hashbrownai/azure';
-import { HashbrownOpenAI } from '@hashbrownai/openai';
 import { HashbrownGoogle } from '@hashbrownai/google';
+import { HashbrownOpenAI } from '@hashbrownai/openai';
 import { HashbrownWriter } from '@hashbrownai/writer';
 import cors from 'cors';
 import 'dotenv/config';
 import express from 'express';
+import OpenAI from 'openai';
 
 const host = process.env.HOST ?? '0.0.0.0';
 const port = process.env.PORT ? Number(process.env.PORT) : 3000;
@@ -80,6 +80,14 @@ app.listen(port, host, () => {
   console.log(`[ ready ] http://${host}:${port}`);
 });
 
+const onCompletion = async (
+  messages: OpenAI.ChatCompletionMessageParam[],
+  completionMessage: Chat.Api.AssistantMessage | null,
+  usage: OpenAI.Completions.CompletionUsage | undefined,
+) => {
+  console.log('onCompletion', messages, completionMessage, usage);
+};
+
 app.post('/chat', async (req, res, next) => {
   const request = req.body as Chat.Api.CompletionCreateParams;
 
@@ -95,6 +103,7 @@ app.post('/chat', async (req, res, next) => {
     stream = HashbrownOpenAI.stream.text({
       apiKey: OPENAI_API_KEY,
       request,
+      onChatCompletion: onCompletion,
     });
   } else if (KNOWN_WRITER_MODEL_NAMES.includes(modelName as KnownModelIds)) {
     stream = HashbrownWriter.stream.text({
